got gnuplot working

+++ src/write.c
@@ -5,10 +5,10 @@
 
 
 
-int WriteTimeData(FILE *fd, int16_t *dp, int numvals) {
+int WriteTimeData(FILE *fd, double *dp, int numvals) {
 	int i;
 	for (i = 0; i < numvals; i++) {
-		fprintf(fd, "%" PRId16 "\n", *(dp+i));
+		fprintf(fd, "%f\n", *(dp+i));
 	}
 	return 0;
 }
