+++ src/write.h
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int WriteTimeData(FILE * fd, int16_t *time_series, int len);
+int WriteTimeData(FILE * fd, double *time_series, int len);
 
 int WriteTimeFreqData(FILE * fd, int16_t *time_series, int16_t * realf, int16_t * imf, int len);
